+++ api/mojang/queue/storage.go
@@ -4,15 +4,15 @@ import "github.com/elyby/chrly/api/mojang"
 
 type UuidsStorage interface {
 	GetUuid(username string) (string, error)
-	StoreUuid(username string, uuid string)
+	StoreUuid(username string, uuid string) error
 }
 
+// nil value can be passed to the storage to indicate that there is no textures
+// for uuid and we know about it. Return err only in case, when storage completely
+// unable to load any information about textures
 type TexturesStorage interface {
-	// nil can be returned to indicate that there is no textures for uuid
-	// and we know about it. Return err only in case, when storage completely
-	// don't know anything about uuid
 	GetTextures(uuid string) (*mojang.SignedTexturesResponse, error)
-	StoreTextures(textures *mojang.SignedTexturesResponse)
+	StoreTextures(uuid string, textures *mojang.SignedTexturesResponse)
 }
 
 type Storage interface {
@@ -31,16 +31,16 @@ func (s *SplittedStorage) GetUuid(username string) (string, error) {
 	return s.UuidsStorage.GetUuid(username)
 }
 
-func (s *SplittedStorage) StoreUuid(username string, uuid string) {
-	s.UuidsStorage.StoreUuid(username, uuid)
+func (s *SplittedStorage) StoreUuid(username string, uuid string) error {
+	return s.UuidsStorage.StoreUuid(username, uuid)
 }
 
 func (s *SplittedStorage) GetTextures(uuid string) (*mojang.SignedTexturesResponse, error) {
 	return s.TexturesStorage.GetTextures(uuid)
 }
 
-func (s *SplittedStorage) StoreTextures(textures *mojang.SignedTexturesResponse) {
-	s.TexturesStorage.StoreTextures(textures)
+func (s *SplittedStorage) StoreTextures(uuid string, textures *mojang.SignedTexturesResponse) {
+	s.TexturesStorage.StoreTextures(uuid, textures)
 }
 
 // This error can be used to indicate, that requested