+++ api/mojang/queue/storage_test.go
@@ -17,8 +17,9 @@ func (m *uuidsStorageMock) GetUuid(username string) (string, error) {
 	return args.String(0), args.Error(1)
 }
 
-func (m *uuidsStorageMock) StoreUuid(username string, uuid string) {
+func (m *uuidsStorageMock) StoreUuid(username string, uuid string) error {
 	m.Called(username, uuid)
+	return nil
 }
 
 type texturesStorageMock struct {
@@ -35,8 +36,8 @@ func (m *texturesStorageMock) GetTextures(uuid string) (*mojang.SignedTexturesRe
 	return result, args.Error(1)
 }
 
-func (m *texturesStorageMock) StoreTextures(textures *mojang.SignedTexturesResponse) {
-	m.Called(textures)
+func (m *texturesStorageMock) StoreTextures(uuid string, textures *mojang.SignedTexturesResponse) {
+	m.Called(uuid, textures)
 }
 
 func TestSplittedStorage(t *testing.T) {
@@ -59,7 +60,7 @@ func TestSplittedStorage(t *testing.T) {
 	t.Run("StoreUuid", func(t *testing.T) {
 		storage, uuidsMock, _ := createMockedStorage()
 		uuidsMock.On("StoreUuid", "username", "result").Once()
-		storage.StoreUuid("username", "result")
+		_ = storage.StoreUuid("username", "result")
 		uuidsMock.AssertExpectations(t)
 	})
 
@@ -74,10 +75,10 @@ func TestSplittedStorage(t *testing.T) {
 	})
 
 	t.Run("StoreTextures", func(t *testing.T) {
-		toStore := &mojang.SignedTexturesResponse{Id: "mock id"}
+		toStore := &mojang.SignedTexturesResponse{}
 		storage, _, texturesMock := createMockedStorage()
-		texturesMock.On("StoreTextures", toStore).Once()
-		storage.StoreTextures(toStore)
+		texturesMock.On("StoreTextures", "mock id", toStore).Once()
+		storage.StoreTextures("mock id", toStore)
 		texturesMock.AssertExpectations(t)
 	})
 }